add userList clone
