+++ 3DAYMarket/public/chatting/chatting.css
@@ -59,7 +59,7 @@ span {
 
 #chat-box {
   width: 100%;
-  display: flex;
+  display: none;
   justify-content: space-between;
   flex-direction: column;
   margin: 0;
@@ -157,11 +157,6 @@ span {
   border-top: 0;
   margin-top: -0.625em;
   margin-left: -1.25em;
-}
-
-li {
-  /* min-height: 5rem; */
-  /* align-items: center; */
 }
 
 h3 {
